requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Supabase
supabase>=2.3.0

# Redis
redis>=5.0.1

# Authentication
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# AI/ML
openai>=1.10.0
anthropic>=0.18.0
cohere>=4.45

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Stripe
stripe>=8.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
python-dotenv>=1.0.0
